+++ ci/travis/cover.sh
@@ -8,14 +8,16 @@ if ! [[ $FEATURES == *"pypy"* ]]; then
   exit 0
 fi
 
-### Run kcov in parallel #######################################################
+### Run kcov ###################################################################
 
 rm -f target/debug/pyo3*.d
 rm -f target/debug/test_*.d
 rm -f target/debug/test_doc-*
 
+# Note: On travis this is run with -P1 because it started failing with
+# `-P $(nproc)`. kcov can probably be run in parallel if used with different CI
 FILES=$(find . -path ./target/debug/pyo3\* -or -path ./target/debug/test_\*)
-echo $FILES | xargs -n1 -P $(nproc) sh -c '
+echo $FILES | xargs -n1 -P1 sh -c '
   dir="target/cov/$(basename $@)"
   mkdir -p $dir
   echo "Collecting coverage data of $(basename $@)"